move routing to line.go
